requirements.txt
torch>=2.1.0
torchvision>=0.16.0
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.27.0
Pillow>=10.0.0
matplotlib>=3.7.0
numpy>=1.24.0
tqdm>=4.65.0
pandas>=2.0.0